Organizing classes, interfaces and tests.
