+++ src/ood/interview/cards/interfaces/Game.java
@@ -1,4 +1,4 @@
-package ood.interview.cards;
+package ood.interview.cards.interfaces;
 
 import java.util.Collection;
